Move all the components
