+++ frontend/src/routes/SingleShowing/ParticipantsList.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-import Header from "./Header";
+import Header from "../../Header";
 
 import { groupBy } from "lodash";
