+++ frontend/src/routes/SingleShowing/SwishModal.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 
 import QRCode from "./QRCode";
-import Loader from "./ProjectorLoader";
+import Loader from "../../ProjectorLoader";
 import Modal from "./Modal";
 
 const SwishModal = ({ payData, closeSwish }) => {
